pyproject.toml
[tool.poetry]
name = "initbot"
version = "0.1.0"
description = "A Discord bot that manages RPG character initiatives"
authors = ["Stefan Götz <github.nooneelse@spamgourmet.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.7.0"
"discord.py" = {version = "^1.7.3", extras = ["voice"]}
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"
types-dataclasses = {version = "^0.1.7", python = "3.6"}
ffmpeg = "^1.4.0"
opuslib = "^3.0.0"
pynacl = "^1.4.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
black = "^21.7b0"
pylint = "^2.9.6"
mypy = "^0.910"
pre-commit = "^2.13.0"
requests = "^2.26.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
